Updated contact form to be cleaner
Since most email clients will thread emails based on sender and subject
line, I'd decided to change the subject line to be more useful.
Moreover, the email body doesn't need to contain the person's name and
email since that is in the reply to field.

The newline characters are not being converted to `<br \>` tags in the
HTML email.