Partner class password protection (#424)
Add feature to password protect class sessions for partnered events.

Once the user has entered the password, it's saved to their account. If the user is not logged in, their browser session is saved so they don't have to keep entering the password. Once they log in (after entering the password), it is saved to their account.