Fixed age calculations. (#489)
The student's age should be calculated compared to the date of a session
or the current date.

On the guardian profile page, the student's age should be calculated in
relation of today's date.

On session or order pages, the age should be calculated based on the
session's start date. This should allow students who will turn a certain
age by the time of a session to be able to sign up.