Add main wrapper to programs template content.
This template overrides the base template's content block so required it's own main wrapper.

This template was already using a main tag, but not wrapping the entire main content of the page.

The existing main tag was changed to a section tag before adding the main wrapper around all main content.