Fix overly broad selector for link style override.
Previously, a CSS selector was defined to override link styling on dark backgrounds (see #812). This looked to be needed
for the display of upcoming class details.

With the changes made in preparation of adding a skip nav link, a main element is now being used to wrap all a page's
main content (for accessibility reasons). Since the previously defined selector was using the main tag at the top of the
selector hierarchy, it is now too broad and matching things that it should not be.

The main tag in that previously defined selector is now replaced with a class selector "class-info-box" instead. This
class selector is now being used in the necessary places, as gleaned from the intent of #812 to achieve the intended
styling.