Add upcoming prereq sessions to view's context.
This adds a dictionary (keyed by Course id) of Session details to the view's context.

These session details are for future sessions being held for prerequisites of the current session.

This detail facilitates adding prerequisite buttons to the registration form when a student is missing one or more of the prerequisites required.