2024 docker fix (#1400)
* chore: Update docker-compose.yml for healthcheck and restart options

This commit updates the docker-compose.yml file to add healthcheck and restart options for the db service. The healthcheck ensures that the PostgreSQL database is ready before starting the app service. The restart option is set to true to automatically restart the db service if it fails.

* chore: Update pyproject.toml, app.json, and Dockerfile

This commit updates the pyproject.toml file to set the package-mode to false. It also updates the app.json file to change the value of the POETRY_VERSION environment variable from 1.4.1 to 1.8.3. Additionally, it updates the Dockerfile to set the value of the POETRY_VERSION environment variable to 1.8.3.

* feat: Update django-allauth version to 0.63.3 and add socialaccount extras

The commit updates the django-allauth dependency from version 0.62.1 to 0.63.3. It also adds the "socialaccount" extras to the django-allauth package.

* chore: Update django_recaptcha dependency and imports

This commit updates the django_recaptcha dependency and replaces the old imports with the new ones. The ReCaptchaField and ReCaptchaV3 classes are now imported from the django_recaptcha package instead of the captcha package.

* chore: Update vscode settings for python code actions

This commit updates the vscode settings for Python code actions. The "source.organizeImports" and "source.fixAll" options are changed from boolean values to "explicit" to ensure explicit control over the code actions.

* chore: Remove "sendgrid" addon from app.json

This commit removes the "sendgrid" addon from the app.json file. The "sendgrid" addon is no longer needed for the project.

* chore: Update Python version to 3.11 in Dockerfile and pyproject.toml

This commit updates the Python version in the Dockerfile and pyproject.toml files from 3.11.3 to ^3.11. This ensures compatibility with the latest Python 3.11.x releases.

* chore: Update Python version to 3.11.9 in Dockerfile and pyproject.toml

* chore: Update PostgreSQL image version to 16-alpine

* feat: Update session start dates to 2025

This commit updates the start dates of the sessions in the `fixtures/14-coderdojochi.session.json` file to the year 2025. The previous start dates were set to 2023. This change ensures that the session data is up to date and reflects the correct start dates for the sessions.

* chore: Update README with up-to-date commands